deploy_package/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow==2.20.0
pillow
numpy
